+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - added store type setting to myems-admin
+- added optional redis database to myems-api
 ### Changed
 ### Fixed
 - optimized combined equipment settings in myems-admin